Moved new group folder to be inside group component folder
