+++ client/components/Group/New/NewGroup.jsx
@@ -5,7 +5,7 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faPlus } from "@fortawesome/free-solid-svg-icons";
 
 // Styling
-import global from "../components.global.scss";
+import global from "../../components.global.scss";
 import styles from "./newgroup.module.scss";
 
 const categories = [
