+++ client/components/Group/New/newgroup.module.scss
@@ -1,4 +1,4 @@
-@import "../../../public/styles/colors.scss";
+@import "../../../../public/styles/colors.scss";
 
 .TermPage {
   margin-top: 48px;
